requirements.txt
alembic==1.11.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
click==8.1.6
coverage==7.3.0
dnspython==2.4.2
email-validator==2.0.0.post2
Faker==19.3.0
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
Flask-WTF==1.1.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
